requirements.txt
bilibili_api>=8.1.0
pillow
youtube-dl
mysql-connector
sqlalchemy
flask
flask-sqlalchemy
feedparser
schedule